refactor - 공통 컴포넌트가 아닌 바텀 시트 컨텐츠 코드 이동
